x/omni/types/message_dewhitelisted_test.go
package types

import (
	"testing"

	"omni/testutil/sample"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgDewhitelisted_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDewhitelisted
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDewhitelisted{
				Authority: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDewhitelisted{
				Authority: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
